Update general information map with felt reports layer
