+++ src/components/maps/GeneralInformation.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-  import { addEpicenter, createMap, id, removeEpicenter, setView } from '@/map_functions'
+  import { addEpicenter, createMap, id, removeEpicenter, setView, addFeltReports } from '@/map_functions'
   import { numberDeclension } from '@/helpers/number'
   import apiSettings from '@/settings/api'
   import security from './mixins/security'
@@ -88,19 +88,7 @@
         })
 
         if (feltReports.length) {
-          const icon = window.L.icon({
-            iconUrl: '/static/img/green-square.png',
-            iconSize: [15, 15]
-          })
-
-          feltReports.forEach((feltReportObject) => {
-            const center = feltReportObject.location
-            const cii = feltReportObject.cii
-            window.L
-              .marker([center.lat, center.lon], {icon})
-              .addTo(this.map.object)
-              .bindPopup(`Интенсивность по ШСИ-2017: ${cii} ${numberDeclension(cii, ['балл', 'балла', 'баллов'])}`)
-          })
+          addFeltReports(this.map.object, feltReports, this.map.object._controls)
         }
 
         // Show map legend just once.
